typescript/rocketseat/back-solid-nodejs/src/usecases/check-in/checkins-use-case.ts
import { CheckIn } from '@prisma/client'
import { ICheckInRepository } from '@/repositories/ICheck-in-repository'
import { IGymsRepository } from '@/repositories/IGyms-repository'
import { ResourceNotFoundError } from '../erros/resource-not-found-error'

interface CheckInUseCaseRequest {
  userId: string
  gymId: string
  userLatitude: number
  userLongitude: number
}

interface CheckInUseCaseResponse {
  checkIn: CheckIn
}

export class CheckInUseCase {
  constructor(
    private readonly checkInRepository: ICheckInRepository,
    private readonly gymRepository: IGymsRepository,
  ) {}

  async execute({
    userId,
    gymId,
  }: CheckInUseCaseRequest): Promise<CheckInUseCaseResponse> {
    const gym = this.gymRepository.findById(gymId)

    if (!gym) {
      throw new ResourceNotFoundError()
    }

    const checkInOnSameDay = await this.checkInRepository.findByUserIdOnDate(
      userId,
      new Date(),
    )

    if (checkInOnSameDay) {
      throw new Error('User already checked in')
    }

    const checkIn = await this.checkInRepository.create({
      user_id: userId,
      gym_id: gymId,
    })

    return {
      checkIn,
    }
  }
}
